refactor: bundle GLSL shaders at build time instead of runtime fetch
Replace runtime shader fetching (from Vercel Blob Storage in prod,
filesystem in dev) with build-time bundling via raw-loader.

This simplifies the architecture by:
- Eliminating dual loading strategy for prod/dev
- Removing runtime async operations for shader loading
- Bundling shaders directly with the app (no external hosting)
- Merging server/client component split into single client component

No functional changes to FlowGradient behavior.
